Add PE351. Cover up a bunch of sloppiness with casts to get rid of noisy compiler conversion warnings.

+++ Problems/PE104.cpp
@@ -29,7 +29,7 @@ uint32_t PE104()
         f2 = f3;
 
         str = f3.get_str();
-        uint32_t nDigits = str.size();
+        uint32_t nDigits = (uint32_t)str.size();
 
         if (nDigits >= 9)
         {
